+++ ecs.php
@@ -1,24 +1,34 @@
-<?php declare(strict_types=1);
+<?php
 
+declare(strict_types=1);
+
+use AlmaOss\CodingStandard\Set\SetList;
 use PhpCsFixer\Fixer\Alias\MbStrFunctionsFixer;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use Symplify\EasyCodingStandard\ValueObject\Option;
+use Symplify\EasyCodingStandard\Config\ECSConfig;
 
-return static function (ContainerConfigurator $containerConfigurator): void {
-    $parameters = $containerConfigurator->parameters();
+$doctrineCacheBridge = [
+    'src/FrontBundle/Adapter/DoctrineCacheBridge.php',
+];
 
-    $parameters->set(
-        Option::SKIP,
-        [
-            MbStrFunctionsFixer::class => [
-                // intentional use of strlen instead of mb_strlen
-                'src/Service/XmlHelper.php',
-                // intentional use of strpos and strlen because of substr_replace is not multi-byte
-                'src/Service/HtmlHelper.php',
-            ],
+return ECSConfig::configure()
+    ->withPaths([
+        __DIR__ . '/src',
+        __DIR__ . '/tests',
+    ])
+    ->withRootFiles()
+    ->withSets([
+        SetList::ALMACAREER,
+    ])
+    ->withSkip([
+        MbStrFunctionsFixer::class => [
+            // intentional use of strlen instead of mb_strlen
+            'src/Service/XmlHelper.php',
+            // intentional use of strpos and strlen because of substr_replace is not multi-byte
+            'src/Service/HtmlHelper.php',
+
+            // using trim, which is not available in 8.1
+            'src/Service/StringUtils.php',
+            'src/Service/LinkPlaceholderTransformer.php',
+            'src/Service/Url.php',
         ],
-    );
-
-    $containerConfigurator->import(__DIR__ . '/tools/coding-standards/vendor/lmc/coding-standard/ecs.php');
-    $containerConfigurator->import(__DIR__ . '/tools/coding-standards/vendor/lmc/coding-standard/ecs-8.1.php');
-};
+    ]);