+++ src/Entity/LetterItem.php
@@ -1,12 +1,12 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace VysokeSkoly\UtilsBundle\Entity;
 
 class LetterItem
 {
-    public function __construct(private string $letter, private array $items = [])
-    {
-    }
+    public function __construct(private string $letter, private array $items = []) {}
 
     public function addItem(mixed $item): void
     {